+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.6.0] - 2020-08-24
+
 ### Added
 
 - `iter.peekable()` inspired by Rust's [`std::iter::Peekable`](https://doc.rust-lang.org/std/iter/struct.Peekable.html).
 
-## Changed
+### Changed
 
 - (docs) Added special note on the inclusivity of `create.range()`.
 
@@ -176,7 +178,7 @@ and this project adheres to
   - `iter.concat` (like `Array.prototype.concat`)
 
 [unreleased]:
-  https://github.com/jajaperson/iterable-utilities/compare/v2.5.0...HEAD
+  https://github.com/jajaperson/iterable-utilities/compare/v2.6.0...HEAD
 [0.2.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.2.0
 [0.1.1]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.1.1
 [0.1.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v0.1.0
@@ -190,3 +192,4 @@ and this project adheres to
 [2.4.1]: https://github.com/jajaperson/iterable-utilities/releases/tag/v2.4.1
 [2.4.2]: https://github.com/jajaperson/iterable-utilities/releases/tag/v2.4.2
 [2.5.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v2.5.0
+[2.6.0]: https://github.com/jajaperson/iterable-utilities/releases/tag/v2.6.0